requirements.txt
PySide6>=6.5.0
pynput>=1.7.6
mss>=9.0.0
Pillow>=10.0.0
Jinja2>=3.1.0
